feat(email): enable react-email
