+++ package.json
@@ -19,7 +19,8 @@
 		"preview": "next build && next start",
 		"start": "next start",
 		"typecheck": "tsc --noEmit",
-		"prepare": "husky"
+		"prepare": "husky",
+		"email:preview": "dotenv -e .env -- email dev --dir ./src/components/emails --port 3030"
 	},
 	"dependencies": {
 		"@auth/prisma-adapter": "^2.7.2",
@@ -28,6 +29,7 @@
 		"@better-upload/server": "^3.0.12",
 		"@prisma/adapter-pg": "^7.2.0",
 		"@prisma/client": "^7.2.0",
+		"@react-email/components": "1.0.4",
 		"@t3-oss/env-nextjs": "^0.12.0",
 		"@tanstack/react-form": "^1.27.7",
 		"@tanstack/react-query": "^5.69.0",
@@ -43,9 +45,9 @@
 		"next": "^16.1.3",
 		"next-themes": "^0.4.6",
 		"pg": "^8.17.1",
-		"react": "^19.0.0",
+		"react": "^19.2.3",
 		"react-day-picker": "^9.13.0",
-		"react-dom": "^19.0.0",
+		"react-dom": "^19.2.3",
 		"react-dropzone": "^14.3.8",
 		"resend": "^6.7.0",
 		"server-only": "^0.0.1",
@@ -58,6 +60,7 @@
 		"@biomejs/biome": "^2.2.5",
 		"@commitlint/cli": "^20.3.1",
 		"@commitlint/config-conventional": "^20.3.1",
+		"@react-email/preview-server": "5.2.5",
 		"@tailwindcss/postcss": "^4.0.15",
 		"@types/node": "^20.14.10",
 		"@types/pg": "^8.16.0",
@@ -68,6 +71,7 @@
 		"lint-staged": "^16.2.7",
 		"postcss": "^8.5.3",
 		"prisma": "^7.2.0",
+		"react-email": "5.2.5",
 		"tailwindcss": "^4.0.15",
 		"tw-animate-css": "^1.4.0",
 		"typescript": "^5.8.2"
